+++ vamp-plugins/dsp/signalconditioning/FiltFilt.cpp
@@ -24,7 +24,7 @@ FiltFilt::FiltFilt( FilterConfig Config )
     m_filtScratchIn = NULL;
     m_filtScratchOut = NULL;
     m_ord = 0;
-	
+
     initialise( Config );
 }
 
@@ -39,7 +39,7 @@ void FiltFilt::initialise( FilterConfig Config )
     m_filterConfig.ord = Config.ord;
     m_filterConfig.ACoeffs = Config.ACoeffs;
     m_filterConfig.BCoeffs = Config.BCoeffs;
-	
+
     m_filter = new Filter( m_filterConfig );
 }
 
@@ -49,8 +49,8 @@ void FiltFilt::deInitialise()
 }
 
 
-void FiltFilt::process(double *src, double *dst, unsigned int length)
-{	
+void FiltFilt::process(float *src, float *dst, unsigned int length)
+{
     unsigned int i;
 
     if (length == 0) return;
@@ -59,19 +59,19 @@ void FiltFilt::process(double *src, double *dst, unsigned int length)
     unsigned int nFact = 3 * ( nFilt - 1);
     unsigned int nExt	= length + 2 * nFact;
 
-    m_filtScratchIn = new double[ nExt ];
-    m_filtScratchOut = new double[ nExt ];
+    m_filtScratchIn = new float[ nExt ];
+    m_filtScratchOut = new float[ nExt ];
 
-	
-    for( i = 0; i< nExt; i++ ) 
+
+    for( i = 0; i< nExt; i++ )
     {
 	m_filtScratchIn[ i ] = 0.0;
 	m_filtScratchOut[ i ] = 0.0;
     }
 
     // Edge transients reflection
-    double sample0 = 2 * src[ 0 ];
-    double sampleN = 2 * src[ length - 1 ];
+    float sample0 = 2 * src[ 0 ];
+    float sampleN = 2 * src[ length - 1 ];
 
     unsigned int index = 0;
     for( i = nFact; i > 0; i-- )
@@ -89,21 +89,21 @@ void FiltFilt::process(double *src, double *dst, unsigned int length)
     {
 	m_filtScratchIn[ i + nFact ] = src[ i ];
     }
-	
+
     ////////////////////////////////
     // Do  0Ph filtering
     m_filter->process( m_filtScratchIn, m_filtScratchOut, nExt);
-	
-    // reverse the series for FILTFILT 
+
+    // reverse the series for FILTFILT
     for ( i = 0; i < nExt; i++)
-    { 
+    {
 	m_filtScratchIn[ i ] = m_filtScratchOut[ nExt - i - 1];
     }
 
-    // do FILTER again 
+    // do FILTER again
     m_filter->process( m_filtScratchIn, m_filtScratchOut, nExt);
-	
-    // reverse the series back 
+
+    // reverse the series back
     for ( i = 0; i < nExt; i++)
     {
 	m_filtScratchIn[ i ] = m_filtScratchOut[ nExt - i - 1 ];
@@ -117,7 +117,7 @@ void FiltFilt::process(double *src, double *dst, unsigned int length)
     for( i = 0; i < length; i++ )
     {
 	dst[ index++ ] = m_filtScratchOut[ i + nFact ];
-    }	
+    }
 
     delete [] m_filtScratchIn;
     delete [] m_filtScratchOut;
