+++ vamp-plugins/dsp/signalconditioning/Framer.cpp
@@ -44,28 +44,28 @@ void Framer::configure( unsigned int frameLength, unsigned int hop )
 
     if( m_dataFrame != NULL )
     {
-	delete [] m_dataFrame;	
+	delete [] m_dataFrame;
 	m_dataFrame = NULL;
     }
-    m_dataFrame = new double[ m_frameLength ];
+    m_dataFrame = new float[ m_frameLength ];
 
     if( m_strideFrame != NULL )
     {
-	delete [] m_strideFrame;	
+	delete [] m_strideFrame;
 	m_strideFrame = NULL;
     }
-    m_strideFrame = new double[ m_stepSize ];
+    m_strideFrame = new float[ m_stepSize ];
 }
 
-void Framer::getFrame(double *dst)
+void Framer::getFrame(float *dst)
 {
 
     if( (m_ulSrcIndex + ( m_frameLength) ) < m_ulSampleLen )
     {
 	for( unsigned int u = 0; u < m_frameLength; u++)
 	{
-	    dst[ u ] = m_srcBuffer[ m_ulSrcIndex++ ]; 
-	}	
+	    dst[ u ] = m_srcBuffer[ m_ulSrcIndex++ ];
+	}
 	m_ulSrcIndex -= ( m_frameLength - m_stepSize );
     }
     else
@@ -77,7 +77,7 @@ void Framer::getFrame(double *dst)
 	{
 	    dst[ u ] = m_srcBuffer[ m_ulSrcIndex++ ];
 	}
-		
+
 	for( unsigned int u = 0; u < zero; u++ )
 	{
 	    dst[ rem + u ] = 0;
@@ -100,10 +100,10 @@ unsigned int Framer::getMaxNoFrames()
     return m_maxFrames;
 }
 
-void Framer::setSource(double *src, unsigned int length)
+void Framer::setSource(float *src, unsigned int length)
 {
     m_srcBuffer = src;
     m_ulSampleLen = length;
 
-    m_maxFrames = (unsigned int)ceil( (double)m_ulSampleLen/(double)m_stepSize ) ;
+    m_maxFrames = (unsigned int)ceil( (float)m_ulSampleLen/(float)m_stepSize ) ;
 }
